Convert loose .cs files to projects
