+++ FPSBooster/FPSBooster.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FPSBooster", "FPSBooster\FPSBooster.csproj", "{CDDAB989-86E0-41E0-9D37-B3BEB63DE8E2}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CDDAB989-86E0-41E0-9D37-B3BEB63DE8E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CDDAB989-86E0-41E0-9D37-B3BEB63DE8E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CDDAB989-86E0-41E0-9D37-B3BEB63DE8E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CDDAB989-86E0-41E0-9D37-B3BEB63DE8E2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {96918ACC-9127-49C2-84D1-A8B03FF7D5E9}
+	EndGlobalSection
+EndGlobal
