+++ FullSatsumaRemover/FullSatsumaRemover.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FullSatsumaRemover", "FullSatsumaRemover\FullSatsumaRemover.csproj", "{9ABA5FC0-26AA-4160-A773-61FE4303D3A9}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9ABA5FC0-26AA-4160-A773-61FE4303D3A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9ABA5FC0-26AA-4160-A773-61FE4303D3A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9ABA5FC0-26AA-4160-A773-61FE4303D3A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9ABA5FC0-26AA-4160-A773-61FE4303D3A9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {271016AB-7B18-4B73-A404-CD1CDA31D0AD}
+	EndGlobalSection
+EndGlobal
