+++ NoMiddleFinger/NoMiddleFinger.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NoMiddleFinger", "NoMiddleFinger\NoMiddleFinger.csproj", "{A617D385-F77C-4F27-AB64-6555E8FF63D4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A617D385-F77C-4F27-AB64-6555E8FF63D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A617D385-F77C-4F27-AB64-6555E8FF63D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A617D385-F77C-4F27-AB64-6555E8FF63D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A617D385-F77C-4F27-AB64-6555E8FF63D4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {503D3579-5AC4-4288-8116-312CF390ED5D}
+	EndGlobalSection
+EndGlobal
