+++ TeimoRemoval/TeimoRemoval.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeimoRemoval", "TeimoRemoval\TeimoRemoval.csproj", "{19BE7D89-553C-47A6-99A0-B2F777223A76}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{19BE7D89-553C-47A6-99A0-B2F777223A76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{19BE7D89-553C-47A6-99A0-B2F777223A76}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{19BE7D89-553C-47A6-99A0-B2F777223A76}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{19BE7D89-553C-47A6-99A0-B2F777223A76}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1A4BB312-34EA-4FFC-BE17-16E1C49E077A}
+	EndGlobalSection
+EndGlobal
