+++ TrafficDisabler/TrafficDisabler.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.10.35013.160
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrafficDisabler", "TrafficDisabler\TrafficDisabler.csproj", "{54A5FAF1-7D1A-4376-887E-88935235B653}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{54A5FAF1-7D1A-4376-887E-88935235B653}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{54A5FAF1-7D1A-4376-887E-88935235B653}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{54A5FAF1-7D1A-4376-887E-88935235B653}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{54A5FAF1-7D1A-4376-887E-88935235B653}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {AE0C19A9-4D98-4624-9277-71D0ED6AB309}
+	EndGlobalSection
+EndGlobal
